# docs/openapi.yml
openapi: 3.0.3
info:
  title: Emoine
  version: '1.0'
  description: Emoine API
paths:
...
  '/presentations':
    get:
      operationId: getPresentations
      summary: プレゼンテーション一覧とその詳細情報を取得
      tags:
        - presentation
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Presentation'

    post:
      operationId: postPresentations
      summary: プレゼンテーションを追加
      tags:
        - presentation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePresentation'
      responses:
        '201':
          description: Created
        '409':
          description: Conflict
...
  '/presentations/{presentationId}':
    parameters:
      - $ref: '#/components/parameters/presentationId'

    get:
      operationId: getPresentation
      summary: プレゼンテーションの詳細情報を取得
      tags:
        - presentation
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Presentation'
        '404':
          description: Not Found

    patch:
      operationId: editPresentation
      summary: プレゼンテーションの詳細情報を変更
      tags:
        - presentation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Presentation'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Presentation'
        '400':
          description: Bad Request
        '404':
          description: Not Found

    delete:
      operationId: deletePresentation
      summary: プレゼンテーションを削除
      tags:
        - presentation
      responses:
        '204':
          description: |-
            No Content
            正常に削除できました。
        '404':
          description: Not Found
...
  '/presentations/{presentationId}/comments':
    parameters:
      - $ref: '#/components/parameters/presentationId'

    get:
      operationId: getPresentationComments
      summary: プレゼンテーションのコメント一覧を取得
      tags:
        - presentation
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
        '404':
          description: Not Found
components:
  parameters:
...
    presentationId:
      name: presentationId
      in: path
      required: true
      description: プレゼンテーションID
      schema:
        type: string
        format: uuid
  schemas:
...
    CreatePresentation:
      type: object
      properties:
        name:
          type: string
        speakers:
          type: string
        description:
          type: string
          minLength: 0
          maxLength: 140
      required:
        - name
        - speakers
        - description
      description: 発表情報
    Presentation:
      type: object
      properties:
        id:
          type: number
          readOnly: true
        name:
          type: string
        speakers:
          type: string
        description:
          type: string
          minLength: 0
          maxLength: 140
        prev:
          type: number
          nullable: true
          description: 前の発表のID
        next:
          type: number
          nullable: true
          description: 次の発表のID
      required:
        - id
        - name
        - speakers
        - description
        - prev
        - next
      description: 発表情報
...
    Comment:
      type: object
      properties:
        id:
          type: number
          readOnly: true
        userId:
          type: string
          format: uuid
          readOnly: true
        presentationId:
          type: number
        text:
          type: string
      required:
        - id
        - userId
        - presentationId
        - text
      description: コメント
...